abstracted client-side code for readability

+++ controllers/stream.coffee
@@ -28,38 +28,41 @@ exports.index = (req, res) ->
 
   http.get(reqOptions(req.params.trackId), (result) ->
     # util.log(util.inspect(result))
-    location = url.parse(result.headers.location)
+    if !result.headers.location
+      res.send "No stream available"
+    else
+      location = url.parse(result.headers.location)
     
-    http.get(optsFromRedirect(location), (result) ->
-      # util.log(util.inspect(result))
-      ffmpeg.createProcessor(
-          inputStream: result
-          outputStream: res
-          emitInputAudioCodecEvent: true
-          emitInfoEvent: true
-          emitProgressEvent: true
-          niceness: 10
-          timeout: 10 * 60 * 1000
-          arguments:
-            "-ab": "128k"
-            "-acodec": "libvorbis"
-            "-f": "ogg"
-            # "-aq": "60"
-        ).on("info", (infoLine) ->
-          util.log infoLine
-        ).on("inputAudioCodec", (codec) ->
-          util.debug "input audio codec is: " + codec
-        ).on("success", (retcode, signal) ->
-          util.debug "process finished successfully with retcode: " + retcode + ", signal: " + signal
-        ).on("failure", (retcode, signal) ->
-          util.debug "process failure, retcode: " + retcode + ", signal: " + signal
-        ).on("progress", (bytes) ->
-          util.debug "process event, bytes: " + bytes
-        ).on("timeout", (processor) ->
-          util.debug "timeout event fired, stopping process."
-          processor.terminate()
-        ).execute() 
-    )
+      http.get(optsFromRedirect(location), (result) ->
+        # util.log(util.inspect(result))
+        ffmpeg.createProcessor(
+            inputStream: result
+            outputStream: res
+            emitInputAudioCodecEvent: true
+            emitInfoEvent: true
+            emitProgressEvent: true
+            niceness: 10
+            timeout: 10 * 60 * 1000
+            arguments:
+              "-ab": "128k"
+              "-acodec": "libvorbis"
+              "-f": "ogg"
+              # "-aq": "60"
+          ).on("info", (infoLine) ->
+            util.log infoLine
+          ).on("inputAudioCodec", (codec) ->
+            util.debug "input audio codec is: " + codec
+          ).on("success", (retcode, signal) ->
+            util.debug "process finished successfully with retcode: " + retcode + ", signal: " + signal
+          ).on("failure", (retcode, signal) ->
+            util.debug "process failure, retcode: " + retcode + ", signal: " + signal
+          ).on("progress", (bytes) ->
+            util.debug "process event, bytes: " + bytes
+          ).on("timeout", (processor) ->
+            util.debug "timeout event fired, stopping process."
+            processor.terminate()
+          ).execute() 
+          res.contentType('application/ogg')
+      )
   )
-  res.contentType('application/ogg')
